+++ flows/definition/localization.go
@@ -1,8 +1,6 @@
 package definition
 
 import (
-	"encoding/json"
-
 	"github.com/nyaruka/gocommon/i18n"
 	"github.com/nyaruka/gocommon/jsonx"
 	"github.com/nyaruka/gocommon/uuids"
@@ -12,16 +10,47 @@ import (
 // holds all property translations for a specific item, e.g.
 //
 //	{
-//	  "text": "Do you like cheese?"
-//	  "quick_replies": ["Yes", "No"]
+//	  "text": ["Do you like cheese?"],
+//	  "quick_replies": ["Yes", "No"],
+//	  "_ui": {...}
 //	}
-type itemTranslation map[string][]string
+type itemTranslation map[string]any
+
+func (t itemTranslation) get(property string) []string {
+	value, found := t[property]
+	if !found {
+		return nil
+	}
+
+	// flow editor is allowed to stuff _ui in here and it's not a string array
+	asSlice, ok := value.([]any)
+	if !ok {
+		return nil
+	}
+
+	trans := make([]string, len(asSlice))
+	for i, v := range asSlice {
+		asString, ok := v.(string)
+		if !ok {
+			return nil
+		}
+		trans[i] = asString
+	}
+
+	// TODO editor sometimes saves empty rule translations as [""] which we should fix in a flow migration
+	// but for now need to ignore
+	if len(trans) == 0 || (len(trans) == 1 && trans[0] == "") {
+		return nil
+	}
+
+	return trans
+}
 
 // holds all the item translations for a specific language, e.g.
 //
 //	{
 //	  "f3368070-8db8-4549-872a-e69a9d060612": {
-//	    "text": "Do you like cheese?"
+//	    "text": ["Do you like cheese?"],
 //	    "quick_replies": ["Yes", "No"]
 //	  },
 //	  "7a1aec43-f3e1-42f0-b967-0ee75e725e3a": { ... }
@@ -32,16 +61,7 @@ type languageTranslation map[uuids.UUID]itemTranslation
 func (t languageTranslation) getTextArray(uuid uuids.UUID, property string) []string {
 	item, found := t[uuid]
 	if found {
-		translation, found := item[property]
-		if found {
-			// TODO editor sometimes saves empty rule translations as [""] which we should fix in a flow migration
-			// but for now need to ignore
-			if len(translation) == 0 || (len(translation) == 1 && translation[0] == "") {
-				return nil
-			}
-
-			return translation
-		}
+		return item.get(property)
 	}
 	return nil
 }
@@ -53,15 +73,12 @@ func (t languageTranslation) setTextArray(uuid uuids.UUID, property string, tran
 		t[uuid] = make(itemTranslation)
 	}
 
-	t[uuid][property] = translated
-}
-
-func (t languageTranslation) Enumerate(callback func(uuids.UUID, string, []string)) {
-	for uuid, it := range t {
-		for property, texts := range it {
-			callback(uuid, property, texts)
-		}
+	trans := make([]any, len(translated))
+	for i, v := range translated {
+		trans[i] = v
 	}
+
+	t[uuid][property] = trans
 }
 
 // our top level container for all the translations for all languages
@@ -100,7 +117,7 @@ func (l localization) SetItemTranslation(lang i18n.Language, itemUUID uuids.UUID
 }
 
 // ReadLocalization reads entire localization flow segment
-func ReadLocalization(data json.RawMessage) (flows.Localization, error) {
+func ReadLocalization(data []byte) (flows.Localization, error) {
 	translations := &localization{}
 	if err := jsonx.Unmarshal(data, translations); err != nil {
 		return nil, err